Several small fixes

+++ UbStudyHelpCore/UbStudyHelp/Pages/UbHelpPage.xaml.cs
@@ -55,7 +55,7 @@ namespace UbStudyHelp.Pages
                 };
 
                 style.Setters.Add(new Setter(Block.FontFamilyProperty, new FontFamily(StaticObjects.Parameters.FontFamily)));
-                style.Setters.Add(new Setter(Block.FontSizeProperty, StaticObjects.Parameters.FontSize));
+                style.Setters.Add(new Setter(Block.FontSizeProperty, (double)StaticObjects.Parameters.FontSize));
                 style.Setters.Add(new Setter(Block.ForegroundProperty, App.Appearance.GetForegroundColorBrush()));
                 return style;
             }
@@ -69,7 +69,7 @@ namespace UbStudyHelp.Pages
             };
 
             style.Setters.Add(new Setter(Block.FontFamilyProperty, new FontFamily(StaticObjects.Parameters.FontFamily)));
-            style.Setters.Add(new Setter(Block.FontSizeProperty, StaticObjects.Parameters.FontSize + size));
+            style.Setters.Add(new Setter(Block.FontSizeProperty, (double)StaticObjects.Parameters.FontSize + size));
             style.Setters.Add(new Setter(Block.ForegroundProperty, App.Appearance.GetHighlightColorBrush()));
             return style;
         }
