+++ UbStudyHelpCore/UbStudyHelp/Printing/UpPrintingPreview.cs
@@ -93,7 +93,7 @@ namespace UbStudyHelp.Printing
             };
 
             style.Setters.Add(new Setter(Control.FontFamilyProperty, new FontFamily(StaticObjects.Parameters.FontFamily)));
-            style.Setters.Add(new Setter(Control.FontSizeProperty, StaticObjects.Parameters.FontSize));
+            style.Setters.Add(new Setter(Control.FontSizeProperty, (double)StaticObjects.Parameters.FontSize));
             if (!(control is ComboBox || control is ListView))
             {
                 style.Setters.Add(new Setter(Control.BackgroundProperty, new SolidColorBrush(Colors.White)));
